ENH: Keep original traceback in DataFrame.apply
When "raise <exception instance>" is used inside a try/except block, the
original stacktrace (containing the code path that raised the original
exception).  This makes debugging difficult since all we know is that an error
occured when `frame._apply_standard` attempted to call the provided function.
Preserve the original stacktrace by using "raise" instead of "raise <exception
instance>".  This facilitates debugginb by allowing us to see where (in the
provided callable) the exception occured.

Added mention of this change in release notes