SQL: revert deprecated tquery/uquery to 0.13 version
As tquery and uquery are deprecated, it is not needed to adapt them to the new sqlalchemy interface + the adapted version changed the signature slightly so it was not fully backwards compatible.
Therefore: put back the old code and removed new implementatation in the classes.
Consequence: the functions are not available for the new sqlalchemy engines, but as they are deprecated this is not needed.