API: remove compat keyword in Categorical constructor
Before the new Categorical work, the default two argument constructor
was expecting "codes and levels" but this was changed to
"values and levels" and a 'compat' kwarg was added, which could
be used to switch to the old style constructor useage.

It was decided that we switch to the new style constructor useage as
the new default (compat=False) and implement a 'from_codes(...)'
constructor.

As the compat codepath is now never triggered and code should be
changed to use the new 'from_codes()' constructor, remove the old
codepath.

Add some warnings if we are pretty sure that the old style
constructor is meant, but unfortunately we don't catch all cases.