BUG: loffset argument not applied for resample().count() on timeseries
closes #12725

The code to do this already existed in the `_downsample` method, which is
called when using functions like `mean`. `max`, etc., but there was
nothing in the `_groupby_and_aggregate` method used for `count` and
`size`. If we pull out the offset code from the `_downsample` method
into a separate method, we can reuse it without duplicating it.

Author: onesandzeroes <onesandzeroes@github.com>

Closes #12757 from onesandzeroes/loffset-count and squashes the following commits:

627429a [onesandzeroes] Add to release notes
fa021bf [onesandzeroes] pep8 style fixes
0c0ca34 [onesandzeroes] Update comment re: issue number
dc79369 [onesandzeroes] Apply loffset when aggregating, e.g. using count()
15aadb9 [onesandzeroes] Extract loffset handling into a general method
70e9765 [onesandzeroes] Add test case for loffset when using count() method