BUG, DEP, DOC: Patch and Align Categorical's Sorting API
Clarifies the meaning of 'sort' in the context of `Categorical` to
mean 'organization' rather than 'order',  as it is possible to call
this method (as well as `sort_values`) when the `Categorical` is
unordered.    Also patches a bug in `Categorical.sort_values` in which
`na_position` was not being respected when  `ascending` was set to
`True`. This commit aligns the behaviour with that of `Series`.
Finally, deprecates `sort` in favor of `sort_values`, which is in
alignment with what was done with `Series` back in #10726.

Closes #12785

Author: gfyoung <gfyoung17@gmail.com>

Closes #12882 from gfyoung/categorical-sort-doc and squashes the following commits:

f324a9c [gfyoung] BUG, DOC, DEP: Patch and Align Categorical's Sorting API