COMPAT: 3.6.1 compat for change in PySlice_GetIndices_Ex
This doesn't actually matter to any tests except for some internal
consistency ones.  Bonus is that it eliminates a warning :<    note
that we aren't actually testing this (yet) on Travis as our 3.6 build
uses conda-forge and 3.6.1 is not there as of yet. Its in defaults
though (and shows up on appveyor build).

Author: Jeff Reback <jeff@reback.net>

Closes #15790 from jreback/py361 and squashes the following commits:

42ddddc [Jeff Reback] change to version < 3
d36902c [Jeff Reback] COMPAT: 3.6.1 compat for change in PySlice_GetIndices_Ex